pyproject.toml
[project]
name = "sphero-vem"
version = "0.1.0"
description = ""
authors = [
    {name = "dv-bt",email = "d.bottone@pm.me"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "torch (>=2.0.0,<2.7.0)",
    "torchvision (>=0.15.0,<0.22.0)",
    "careamics (>=0.0.9)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "kornia (>=0.8.0,<0.9.0)",
    "ome-types (>=0.6.0,<0.7.0)",
    "h5py (>=3.13.0,<4.0.0)",
    "lightning (>=2.5.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "cellpose (>=4.0.3,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "simpleitk (>=2.5.2,<3.0.0)",
    "jsonargparse (>=4.40.2,<5.0.0)",
    "tyro (>=0.9.31,<0.10.0)",
    "dask-image (>=2025.11.0,<2026.0.0)",
]

[tool.poetry]
packages = [{include = "sphero_vem", from = "src"}]


[tool.poetry.group.notebook.dependencies]
ipykernel = "^6.29.5"
ipympl = "^0.9.7"
seaborn = "^0.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
ruff = "^0.11.4"
pre-commit = "^4.2.0"
wandb = "^0.19.10"


[tool.poetry.group.gpu.dependencies]
cupy-cuda12x = "^13.6.0"
cucim = "^23.10.0"
dask-cuda = {extras = ["cu12"], version = "^25.10.0"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
